Add new functionalities to the GUI explorer library.
Add the option to limit the clip on up and down edge in case if we want to display statuses and other data that are displayed independently of the GUI explorer.
Add the feature that the selected file will have displayed an asterisc.
Add call's for left and right button's press.